# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "FieldIn": {
        "properties": {
          "entity": {
            "enum": [
              "customer",
              "appointment"
            ],
            "title": "Entity",
            "type": "string"
          },
          "key": {
            "pattern": "^[a-z][a-z0-9_]{0,62}$",
            "title": "Key",
            "type": "string"
          },
          "label_i18n": {
            "additionalProperties": {
              "maxLength": 120,
              "minLength": 1,
              "type": "string"
            },
            "minProperties": 1,
            "title": "Label I18N",
            "type": "object"
          },
          "options": {
            "default": [],
            "items": {
              "maxLength": 120,
              "minLength": 1,
              "type": "string"
            },
            "title": "Options",
            "type": "array"
          },
          "pii_level": {
            "default": "personal",
            "enum": [
              "none",
              "personal",
              "sensitive"
            ],
            "title": "Pii Level",
            "type": "string"
          },
          "position": {
            "default": 0,
            "title": "Position",
            "type": "integer"
          },
          "type": {
            "enum": [
              "text",
              "long_text",
              "number",
              "date",
              "boolean",
              "select",
              "phone",
              "email"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "entity",
          "key",
          "label_i18n",
          "type"
        ],
        "title": "FieldIn",
        "type": "object"
      },
      "FieldOut": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "entity": {
            "enum": [
              "customer",
              "appointment"
            ],
            "title": "Entity",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label_i18n": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Label I18N",
            "type": "object"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "title": "Options",
            "type": "array"
          },
          "pii_level": {
            "enum": [
              "none",
              "personal",
              "sensitive"
            ],
            "title": "Pii Level",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "type": {
            "enum": [
              "text",
              "long_text",
              "number",
              "date",
              "boolean",
              "select",
              "phone",
              "email"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "entity",
          "key",
          "label_i18n",
          "type",
          "options",
          "pii_level",
          "position",
          "archived"
        ],
        "title": "FieldOut",
        "type": "object"
      },
      "FieldPatch": {
        "properties": {
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          },
          "label_i18n": {
            "anyOf": [
              {
                "additionalProperties": {
                  "maxLength": 120,
                  "minLength": 1,
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label I18N"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 120,
                  "minLength": 1,
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "pii_level": {
            "anyOf": [
              {
                "enum": [
                  "none",
                  "personal",
                  "sensitive"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pii Level"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          }
        },
        "title": "FieldPatch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Interval": {
        "properties": {
          "end": {
            "format": "time",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "Interval",
        "type": "object"
      },
...
      "LocationIn": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "opening_hours": {
            "$ref": "#/components/schemas/WeeklyHours",
            "default": {
              "fri": [],
              "mon": [],
              "sat": [],
              "sun": [],
              "thu": [],
              "tue": [],
              "wed": []
            }
          },
          "timezone": {
            "default": "Europe/Brussels",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "LocationIn",
        "type": "object"
      },
      "LocationOut": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "opening_hours": {
            "$ref": "#/components/schemas/WeeklyHours"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "timezone",
          "opening_hours"
        ],
        "title": "LocationOut",
        "type": "object"
      },
      "LocationPatch": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "opening_hours": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeeklyHours"
              },
              {
                "type": "null"
              }
            ]
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "LocationPatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WeeklyHours": {
        "description": "Opening hours per weekday; a missing or empty day is closed.",
        "properties": {
          "fri": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "title": "Fri",
            "type": "array"
          },
          "mon": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "title": "Mon",
            "type": "array"
          },
          "sat": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "title": "Sat",
            "type": "array"
          },
          "sun": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "title": "Sun",
            "type": "array"
          },
          "thu": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "title": "Thu",
            "type": "array"
          },
          "tue": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "title": "Tue",
            "type": "array"
          },
          "wed": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "title": "Wed",
            "type": "array"
          }
        },
        "title": "WeeklyHours",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Confluo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/fields": {
      "get": {
        "operationId": "listFields",
        "parameters": [
          {
            "in": "query",
            "name": "entity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "customer",
                    "appointment"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FieldOut"
                  },
                  "title": "Response Listfields",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Fields",
        "tags": [
          "fields"
        ]
      },
      "post": {
        "operationId": "createField",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Field",
        "tags": [
          "fields"
        ]
      }
    },
    "/api/fields/{field_id}": {
      "delete": {
        "description": "Archive rather than delete: past appointments keep their values.",
        "operationId": "archiveField",
        "parameters": [
          {
            "in": "path",
            "name": "field_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Field Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Field",
        "tags": [
          "fields"
        ]
      },
      "patch": {
        "operationId": "updateField",
        "parameters": [
          {
            "in": "path",
            "name": "field_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Field Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Field",
        "tags": [
          "fields"
        ]
      }
    },
...
    "/api/locations": {
      "get": {
        "operationId": "listLocations",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LocationOut"
                  },
                  "title": "Response Listlocations",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Locations",
        "tags": [
          "locations"
        ]
      },
      "post": {
        "operationId": "createLocation",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Location",
        "tags": [
          "locations"
        ]
      }
    },
    "/api/locations/{location_id}": {
      "delete": {
        "operationId": "deleteLocation",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Location",
        "tags": [
          "locations"
        ]
      },
      "patch": {
        "operationId": "updateLocation",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Tenant-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Location",
        "tags": [
          "locations"
        ]
      }
    },
...
  }
}